pyproject.toml
[project]
name = "sagemaker-mlops-ref-arch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3-stubs>=1.42.79",
    "botocore[crt]>=1.42.59",
    "dagster==1.12.17",
    "dagster-aws>=0.28.17",
    "ipykernel>=7.2.0",
    "nbconvert>=7.17.0",
    "numpy>=2.4.2",
    "pyspark==3.5.2",
    "sagemaker-feature-store-pyspark-3-1>=1.1.3",
    "sagemaker[mlops,train]>=3.7.0",
]
